Travel_Together_Backend/requirements.txt
# Core
Django==4.2.16
djangorestframework==3.15.2
django-environ==0.11.2
django-cors-headers==4.4.0

# Auth
djangorestframework-simplejwt==5.3.1
social-auth-app-django==5.4.2
bcrypt==4.2.0
firebase-admin==6.5.0

# Database
psycopg2-binary==2.9.9

# GeoDjango (PostGIS)
# GDAL, GEOS, PROJ come from the PostGIS Windows install
# gdal is referenced via GDAL_LIBRARY_PATH in settings

# Real-time
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2

# Celery
celery==5.4.0
django-celery-beat==2.7.0
redis==5.0.8

# Storage & Media
Pillow==10.4.0
django-storages==1.14.4
boto3==1.35.0

# Utilities
django-filter==24.3
PyJWT[crypto]==2.11.0
